+++ app/mailers/message_mailer.rb
@@ -15,7 +15,7 @@ class MessageMailer < ApplicationMailer
     mail(
       to: @recipient.email,
       subject: @notification.email_subject,
-      reply_to: "message-#{conversation_token}@inbound.railsdevs.com"
+      reply_to: "conversation+#{conversation_token}@inbound.railsdevs.com"
     )
   end
 end